requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
defusedxml==0.7.1
faiss-cpu==1.9.0.post1
fastapi==0.115.7
ffmpeg==1.4
ffprobe==0.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kagglehub==0.3.6
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.25.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.14
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.48.1
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
yt-dlp==2025.1.15
